% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SATS.R
\name{context4signeR}
\alias{context4signeR}
\title{generate a context matrix for SATS}
\usage{
context4signeR(context)
}
\arguments{
\item{context}{context R object getting from OCA Plus file.}
}
\value{
A matirx.
}
\description{
generate a context matrix for SATS
}
\examples{
\dontrun{
subsCount(context)
}
}
