% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SATS.R
\name{cosmicSignature}
\alias{cosmicSignature}
\title{generate a tibble of  Cosmic TMB signature}
\usage{
cosmicSignature(data4signeR, data4sats)
}
\arguments{
\item{data4signeR}{data for signeR}

\item{data4sats}{data for signeR}
}
\value{
A tiblle
}
\description{
generate a tibble of  Cosmic TMB signature
}
\examples{
\dontrun{
cosmicSignature(data)
}
}
