% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHandle.R
\name{extractID}
\alias{extractID}
\title{generate a context matrix for SATS}
\usage{
extractID(fileList)
}
\arguments{
\item{fileList}{file list generated from fileContext.}
}
\value{
a character vector
}
\description{
generate a context matrix for SATS
}
\examples{
\dontrun{
extractID(fileList)
}
}
