% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileHandle.R
\name{filesContext}
\alias{filesContext}
\title{generate a file list}
\usage{
filesContext(filePath)
}
\arguments{
\item{filePath}{context R object getting from OCA Plus file.}
}
\value{
a character vector
}
\description{
generate a file list
}
\examples{
\dontrun{
filesContext(here::here("path/to/contextfiles"))
}
}
