% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SATS.R
\name{mmrdSignature}
\alias{mmrdSignature}
\title{generate a tibble of  MMRd TMB signature}
\usage{
mmrdSignature(data4signeR, data4sats)
}
\arguments{
\item{data4signeR}{data for signeR}

\item{data4sats}{data for signeR}
}
\value{
A tiblle
}
\description{
generate a tibble of  MMRd TMB signature
}
\examples{
\dontrun{
mmrdSignature(data)
}
}
