% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFile.R
\name{renameFile}
\alias{renameFile}
\title{Fix a filename consistantly}
\usage{
renameFile(filename)
}
\arguments{
\item{filename}{context R object getting from OCA Plus file.}
}
\value{
A file name characters
}
\description{
Fix a filename consistantly
}
\examples{
\dontrun{
renameFile(filename)
}
}
